+++ buildpacks/nodejs-engine/CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Added node version 14.20.1, 18.9.1, 16.17.1.
+- Upgrade `libcnb` and `libherokubuildpack` to `0.11.0`. ([#360](https://github.com/heroku/buildpacks-nodejs/pull/360))
+
 ## [0.8.8] 2022/09/12
 
 - Added node version 18.9.0.